+++ webviz_subsurface/plugins/_parameter_analysis/views/main_view.py
@@ -1,27 +1,43 @@
+from typing import Callable
+
 import dash_html_components as html
 import dash_core_components as dcc
+from webviz_config import WebvizConfigTheme
 
 from .parameter_qc_view import parameter_qc_view
 from .parameter_response_view import parameter_response_view
+from ..models import ParametersModel, SimulationTimeSeriesModel
 
 
-def main_view(parent) -> dcc.Tabs:
+def main_view(
+    get_uuid: Callable,
+    vectormodel: SimulationTimeSeriesModel,
+    parametermodel: ParametersModel,
+    theme: WebvizConfigTheme,
+) -> dcc.Tabs:
     tabs = [
         make_tab(
             label="Parameter distributions",
-            children=parameter_qc_view(parent=parent),
+            children=parameter_qc_view(
+                get_uuid=get_uuid, parametermodel=parametermodel
+            ),
         )
     ]
-    if parent.vmodel is not None:
+    if vectormodel is not None:
         tabs.append(
             make_tab(
                 label="Parameters impact on simulation profiles",
-                children=parameter_response_view(parent=parent),
+                children=parameter_response_view(
+                    get_uuid=get_uuid,
+                    parametermodel=parametermodel,
+                    vectormodel=vectormodel,
+                    theme=theme,
+                ),
             )
         )
 
     return html.Div(
-        id=parent.uuid("layout"),
+        id=get_uuid("layout"),
         children=dcc.Tabs(
             style={"width": "100%"},
             persistence=True,